+++ crates/proto/src/domain/account.rs
@@ -6,8 +6,10 @@ use miden_objects::{
     account::{Account, AccountHeader, AccountId},
     block::{AccountWitness, BlockNumber},
     crypto::hash::rpo::RpoDigest,
+    note::{NoteExecutionMode, NoteTag},
     utils::{Deserializable, DeserializationError, Serializable},
 };
+use thiserror::Error;
 
 use super::try_convert;
 use crate::{
@@ -278,4 +280,77 @@ impl TryFrom<proto::responses::AccountTransactionInputRecord> for AccountState {
 
         Ok(Self { account_id, account_commitment })
     }
+}
+
+// NETWORK ACCOUNT PREFIX
+// ================================================================================================
+
+pub type AccountPrefix = u32;
+
+/// Newtype wrapper for network account prefix.
+/// Provides type safety for accounts that are meant for network execution.
+pub struct NetworkAccountPrefix(u32);
+
+impl NetworkAccountPrefix {
+    pub fn inner(&self) -> u32 {
+        self.0
+    }
+}
+
+impl From<NetworkAccountPrefix> for u32 {
+    fn from(value: NetworkAccountPrefix) -> Self {
+        value.inner()
+    }
+}
+
+impl TryFrom<u32> for NetworkAccountPrefix {
+    type Error = NetworkAccountError;
+
+    fn try_from(value: u32) -> Result<Self, Self::Error> {
+        if value >> 30 != 0 {
+            return Err(NetworkAccountError::InvalidPrefix(value));
+        }
+        Ok(NetworkAccountPrefix(value))
+    }
+}
+
+impl TryFrom<AccountId> for NetworkAccountPrefix {
+    type Error = NetworkAccountError;
+
+    fn try_from(id: AccountId) -> Result<Self, Self::Error> {
+        if !id.is_network() {
+            return Err(NetworkAccountError::NotNetworkAccount(id));
+        }
+        let prefix = get_account_id_tag_prefix(id);
+        Ok(NetworkAccountPrefix(prefix))
+    }
+}
+
+impl TryFrom<NoteTag> for NetworkAccountPrefix {
+    type Error = NetworkAccountError;
+
+    fn try_from(tag: NoteTag) -> Result<Self, Self::Error> {
+        if tag.execution_mode() != NoteExecutionMode::Network || !tag.is_single_target() {
+            return Err(NetworkAccountError::InvalidExecutionMode(tag));
+        }
+
+        let tag_inner: u32 = tag.into();
+        assert!(tag_inner >> 30 == 0, "first 2 bits have to be 0");
+        Ok(NetworkAccountPrefix(tag_inner))
+    }
+}
+
+#[derive(Debug, Error)]
+pub enum NetworkAccountError {
+    #[error("account ID {0} is not a valid network account ID")]
+    NotNetworkAccount(AccountId),
+    #[error("note tag {0} is not valid for network account execution")]
+    InvalidExecutionMode(NoteTag),
+    #[error("note prefix should be 30-bit long ({0} has non-zero in the 2 most significant bits)")]
+    InvalidPrefix(u32),
+}
+
+/// Gets the 30-bit prefix of the account ID.
+fn get_account_id_tag_prefix(id: AccountId) -> AccountPrefix {
+    (id.prefix().as_u64() >> 34) as AccountPrefix
 }