+++ crates/proto/src/domain/note.rs
@@ -1,11 +1,12 @@
 use miden_objects::{
     Digest, Felt,
     note::{
-        Note, NoteDetails, NoteExecutionHint, NoteId, NoteInclusionProof, NoteMetadata, NoteTag,
-        NoteType,
+        Note, NoteDetails, NoteExecutionHint, NoteExecutionMode, NoteId, NoteInclusionProof,
+        NoteMetadata, NoteTag, NoteType, Nullifier,
     },
-    utils::Serializable,
+    utils::{Deserializable, Serializable},
 };
+use thiserror::Error;
 
 use crate::{
     errors::{ConversionError, MissingFieldHelper},
@@ -95,5 +96,72 @@ impl TryFrom<&proto::NoteInclusionInBlockProof> for (NoteId, NoteInclusionProof)
                     .try_into()?,
             )?,
         ))
+    }
+}
+
+// NETWORK NOTE
+// ================================================================================================
+
+/// A newtype that wraps around notes targeting a single account to be used in a network mode.
+#[derive(Clone, Debug, PartialEq, Eq)]
+pub struct NetworkNote(Note);
+
+impl NetworkNote {
+    pub fn inner(&self) -> &Note {
+        &self.0
+    }
+
+    pub fn metadata(&self) -> &NoteMetadata {
+        self.inner().metadata()
+    }
+
+    pub fn nullifier(&self) -> Nullifier {
+        self.inner().nullifier()
+    }
+
+    pub fn id(&self) -> NoteId {
+        self.inner().id()
+    }
+}
+
+impl From<NetworkNote> for Note {
+    fn from(value: NetworkNote) -> Self {
+        value.0
+    }
+}
+
+impl TryFrom<Note> for NetworkNote {
+    type Error = NetworkNoteError;
+
+    fn try_from(note: Note) -> Result<Self, Self::Error> {
+        if !note.metadata().tag().is_single_target()
+            || note.metadata().tag().execution_mode() != NoteExecutionMode::Network
+        {
+            return Ok(NetworkNote(note));
+        }
+        Err(NetworkNoteError::InvalidExecutionMode(note.metadata().tag()))
+    }
+}
+
+#[derive(Debug, Error)]
+pub enum NetworkNoteError {
+    #[error("note tag {0} is not a valid network note tag")]
+    InvalidExecutionMode(NoteTag),
+}
+
+impl TryFrom<proto::NetworkNote> for NetworkNote {
+    type Error = ConversionError;
+
+    fn try_from(proto_note: proto::NetworkNote) -> Result<Self, Self::Error> {
+        let details = NoteDetails::read_from_bytes(&proto_note.details)
+            .map_err(|err| ConversionError::deserialization_error("NoteDetails", err))?;
+        let (assets, recipient) = details.into_parts();
+        let metadata: NoteMetadata = proto_note
+            .metadata
+            .ok_or_else(|| proto::NetworkNote::missing_field(stringify!(metadata)))?
+            .try_into()?;
+        let note = Note::new(assets, metadata, recipient);
+
+        Ok(NetworkNote::try_from(note)?)
     }
 }